pyproject.toml
[tool.poetry]
name = "pyjtalk"
version = "0.2.0"
description = "OpenJTalk Wrapper for Python"
authors = ["alcnaka <47954083+alcnaka@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
